Changes as per comments in the PR
