+++ src/runtime_src/xdp/profile/plugin/aie_trace/util/aie_trace_config.cpp
@@ -462,4 +462,86 @@ namespace xdp::aie::trace {
     return true;
   }
 
+  /****************************************************************************
+   * Reset timer for the specified tile range
+   ***************************************************************************/
+  void timerSyncronization(XAie_DevInst* aieDevInst, xaiefal::XAieDev* aieDevice, std::shared_ptr<AieTraceMetadata> metadata, uint8_t startCol, uint8_t numCols)
+  {
+    std::shared_ptr<xaiefal::XAieBroadcast> traceStartBroadcastCh1 = nullptr, traceStartBroadcastCh2 = nullptr;
+    std::vector<XAie_LocType> vL;
+    traceStartBroadcastCh1 = aieDevice->broadcast(vL, XAIE_PL_MOD, XAIE_CORE_MOD);
+    traceStartBroadcastCh1->reserve();
+    traceStartBroadcastCh2 = aieDevice->broadcast(vL, XAIE_PL_MOD, XAIE_CORE_MOD);
+    traceStartBroadcastCh2->reserve();
+
+    uint8_t broadcastId1 = traceStartBroadcastCh1->getBc();
+    uint8_t broadcastId2 = traceStartBroadcastCh2->getBc();
+
+    //build broadcast network
+    aie::trace::build2ChannelBroadcastNetwork(aieDevInst, metadata, broadcastId1, broadcastId2, XAIE_EVENT_USER_EVENT_0_PL, startCol, numCols);
+
+    //set timer control register
+    for (auto& tileMetric : metadata->getConfigMetrics()) {
+      auto tile       = tileMetric.first;
+      auto col        = tile.col + startCol; 
+      auto row        = tile.row;
+      auto type       = aie::getModuleType(row, metadata->getRowOffset());
+      auto loc        = XAie_TileLoc(col, row);
+
+      if(type == module_type::shim) {
+        XAie_Events resetEvent = (XAie_Events)(XAIE_EVENT_BROADCAST_A_0_PL + broadcastId2);
+        if(col == startCol)
+        {
+          resetEvent = XAIE_EVENT_USER_EVENT_0_PL;
+        }
+
+        XAie_SetTimerResetEvent(aieDevInst, loc, XAIE_PL_MOD, resetEvent, XAIE_RESETDISABLE);
+      }
+      else if(type == module_type::mem_tile) {
+        XAie_Events resetEvent = (XAie_Events) (XAIE_EVENT_BROADCAST_0_MEM_TILE + broadcastId1);
+        XAie_SetTimerResetEvent(aieDevInst, loc, XAIE_MEM_MOD, resetEvent, XAIE_RESETDISABLE);
+      }
+      else {
+        XAie_Events resetEvent = (XAie_Events) (XAIE_EVENT_BROADCAST_0_CORE + broadcastId1);
+        XAie_SetTimerResetEvent(aieDevInst, loc, XAIE_CORE_MOD, resetEvent, XAIE_RESETDISABLE);
+        resetEvent = (XAie_Events) (XAIE_EVENT_BROADCAST_0_MEM + broadcastId1);
+        XAie_SetTimerResetEvent(aieDevInst, loc, XAIE_MEM_MOD, resetEvent, XAIE_RESETDISABLE);
+      }
+    }
+
+    //Generate the event to trigger broadcast network to reset timer
+    XAie_EventGenerate(aieDevInst, XAie_TileLoc(startCol, 0), XAIE_PL_MOD, XAIE_EVENT_USER_EVENT_0_PL);
+
+    //reset timer control register so that timer are not reset after this point
+    for (auto& tileMetric : metadata->getConfigMetrics()) {
+      auto tile       = tileMetric.first;
+      auto col        = tile.col + startCol;
+      auto row        = tile.row;
+      auto type       = aie::getModuleType(row, metadata->getRowOffset());
+      auto loc        = XAie_TileLoc(col, row);
+
+      if(type == module_type::shim) {
+        XAie_Events resetEvent = XAIE_EVENT_NONE_PL ;
+        XAie_SetTimerResetEvent(aieDevInst, loc, XAIE_PL_MOD, resetEvent, XAIE_RESETDISABLE);
+      }
+      else if(type == module_type::mem_tile) {
+        XAie_Events resetEvent = XAIE_EVENT_NONE_MEM_TILE;
+        XAie_SetTimerResetEvent(aieDevInst, loc, XAIE_MEM_MOD, resetEvent, XAIE_RESETDISABLE);
+      }
+      else {
+        XAie_Events resetEvent = XAIE_EVENT_NONE_CORE;
+        XAie_SetTimerResetEvent(aieDevInst, loc, XAIE_CORE_MOD, resetEvent, XAIE_RESETDISABLE);
+        resetEvent = XAIE_EVENT_NONE_MEM;
+        XAie_SetTimerResetEvent(aieDevInst, loc, XAIE_MEM_MOD, resetEvent, XAIE_RESETDISABLE);
+      }
+    }
+
+    //reset broadcast network
+    reset2ChannelBroadcastNetwork(aieDevInst, metadata, broadcastId1, broadcastId2, startCol, numCols);
+
+    //release the channels used for timer sync
+    traceStartBroadcastCh1->release();
+    traceStartBroadcastCh2->release();
+  }
+
 }  // namespace xdp
