+++ src/runtime_src/xdp/profile/plugin/aie_trace/util/aie_trace_config.h
@@ -20,6 +20,7 @@
 #include <cstdint>
 #include "xaiefal/xaiefal.hpp"
 #include "xdp/profile/database/static_info/aie_constructs.h"
+#include "xdp/profile/plugin/aie_trace/aie_trace_metadata.h"
 
 namespace xdp::aie::trace {  
   /**
@@ -119,6 +120,18 @@ namespace xdp::aie::trace {
    */
   bool configStartIteration(xaiefal::XAieMod& core, uint32_t iteration,
                             XAie_Events& startEvent);
+  
+  /**
+   * @brief Reset timers for specified tile range
+   * @param aieDevInst     AIE device Instance
+   * @param aieDevice      AIE device
+   * @param metadata       Trace Metadata
+   * @param startCol       Start column of the partition
+   * @param numCols        Num of columns in the partition
+   */
+  void timerSyncronization(XAie_DevInst* aieDevInst, xaiefal::XAieDev* aieDevice,
+                           std::shared_ptr<AieTraceMetadata> metadata, uint8_t startCol, 
+                           uint8_t numCols);
 }  // namespace xdp::aie::trace
 
 #endif
