+++ src/runtime_src/xdp/profile/plugin/aie_trace/ve2/aie_trace.h
@@ -22,7 +22,6 @@
 #include "xaiefal/xaiefal.hpp"
 #include "xdp/profile/plugin/aie_trace/aie_trace_impl.h"
 #include "xdp/profile/plugin/aie_trace/util/aie_trace_config.h"
-#include "xdp/profile/plugin/aie_trace/util/aie_trace_common_config.h"
 
 namespace xdp {
